feat: allow custom allocator
Adds the ability to use a custom allocator instead of the Global
allocator to create page tables.
This is especially helpful when being on an embedded system that
has not yet been bootstrapped to support the Global locking due to
missing atomic instructions.
For instance, in ARM-v8 (and others), you cannot use atomic instructions
unless having the MMU set up, which creates a chicken or the egg
problem.